Support advert modal display variations
Sometimes the advert modal presented on the Fixtures website is present
in the DOM but not visible due to a display:none; CSS tag.

They may be varying the presence of the advertisement based on timing or
other factors.

The code should now support both cases, though will need the modal to be
displayed again to fully end-to-end test.